Init View with jquery crud actions, fix controller

+++ WebDemoApi/Controllers/JapaneseWordController.cs
@@ -14,39 +14,58 @@ namespace WebDemoApi.Controllers
     public class JapaneseWordController : ApiController
     {
         // GET: api/JapaneseWord
-        public string Get()
+        public IEnumerable<JapaneseWord> Get()
         {
             var jprepository = new JapaneseWordRepository();
 
-            return JsonConvert.SerializeObject(jprepository.GetAllEntries());
+            return jprepository.GetAllEntries();
         }
 
         // GET: api/JapaneseWord/5
-        public string Get(int id)
+        public JapaneseWord Get(int id)
         {
             var jprepository = new JapaneseWordRepository();
-            return JsonConvert.SerializeObject(jprepository.GetEntry(id));
+            return jprepository.GetEntry(id);
         }
 
         // POST: api/JapaneseWord
-        public void Post([FromBody]string value)
+        public HttpResponseMessage Post(JapaneseWord model)
         {
-            var model = new JapaneseWord();
-            // fill model??
+            
             var jprepository = new JapaneseWordRepository();
             jprepository.AddEntry(model);
+
+            var response = Request.CreateResponse<JapaneseWord>(HttpStatusCode.Created, model);
+
+            string uri = Url.Link("DefaultApi", new { id = model.EntryID });
+            response.Headers.Location = new Uri(uri);
+            return response;
 
         }
 
         // PUT: api/JapaneseWord/5
-        public void Put(int id, [FromBody]string value)
+        public void Put(int id, JapaneseWord model)
         {
+            var jprepository = new JapaneseWordRepository();
+            try 
+            {
+                jprepository.EditEntry(model);
+            }
+            catch 
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound);
+            }
+            
         }
 
         // DELETE: api/JapaneseWord/5
         public void Delete(int id)
         {
             var jprepository = new JapaneseWordRepository();
+            if (jprepository.GetEntry(id) == null) 
+            {
+                throw new HttpResponseException(HttpStatusCode.NotFound);
+            }
             jprepository.DeleteEntry(id);
         }
     }
